Process salary

+++ PayrollManager/models.py
@@ -61,13 +61,13 @@ class Salary(models.Model):
     month = models.IntegerField()  # Tháng tính lương (1-12)
     year = models.IntegerField()  # Năm tính lương
     # basic_days = models.IntegerField()  # Số ngày cơ bản
-    basic_days_after_holidays = models.IntegerField()  # Số ngày cơ bản sau lễ
+    basic_days_after_holidays = models.FloatField()  # Số ngày cơ bản sau lễ
     # basic_hours = models.FloatField()  # Số giờ cơ bản
     basic_hours_after_holidays = models.FloatField()  # Số giờ cơ bản sau lễ
     actual_work_hours = models.FloatField()  # Số giờ làm việc thực tế
-    worked_days = models.IntegerField()  # Số ngày đi làm
+    worked_days = models.FloatField()  # Số ngày đi làm
     penalty_hours = models.FloatField()  # Số giờ vi phạm
-    worked_day_off_days = models.IntegerField()  # Số ngày nghỉ đi làm
+    worked_day_off_days = models.FloatField()  # Số ngày nghỉ đi làm
     sunday_hours = models.FloatField()  # Số giờ làm vào Chủ nhật
     holiday_hours = models.FloatField()  # Số giờ lễ
     worked_holiday_hours = models.FloatField()  # Số giờ làm ngày lễ
